[UPDATE] erands and expose layerg config

+++ console/console.pb.gw.go
@@ -1205,7 +1205,7 @@ func request_Console_AddNFTCollection_0(ctx context.Context, marshaler runtime.M
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "chain_id")
 	}
 
-	protoReq.ChainId, err = runtime.Uint32(val)
+	protoReq.ChainId, err = runtime.Uint64(val)
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "chain_id", err)
 	}
@@ -1261,7 +1261,7 @@ func local_request_Console_AddNFTCollection_0(ctx context.Context, marshaler run
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "chain_id")
 	}
 
-	protoReq.ChainId, err = runtime.Uint32(val)
+	protoReq.ChainId, err = runtime.Uint64(val)
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "chain_id", err)
 	}
